design : stroke text 수정
