+++ Outline/Outline Watch/Source/DesignSystem/Component/ElapsedTimeFormatter.swift
@@ -1,0 +1,36 @@
+//
+//  File.swift
+//  Outline Watch App
+//
+//  Created by 김하은 on 10/19/23.
+//
+
+import Foundation
+
+class ElapsedTimeFormatter: Formatter {
+    let componentsFormatter: DateComponentsFormatter = {
+        let formatter = DateComponentsFormatter()
+        formatter.allowedUnits = [.minute, .second]
+        formatter.zeroFormattingBehavior = .pad
+        return formatter
+    }()
+    var showSubseconds = true
+
+    override func string(for value: Any?) -> String? {
+        guard let time = value as? TimeInterval else {
+            return nil
+        }
+
+        guard let formattedString = componentsFormatter.string(from: time) else {
+            return nil
+        }
+
+        if showSubseconds {
+            let hundredths = Int((time.truncatingRemainder(dividingBy: 1)) * 100)
+            let decimalSeparator = Locale.current.decimalSeparator ?? "."
+            return String(format: "%@%@%0.2d", formattedString, decimalSeparator, hundredths)
+        }
+
+        return formattedString
+    }
+}
