+++ Outline/Outline Watch/Source/View/ControlsView.swift
@@ -71,9 +71,8 @@ struct ControlsView: View {
                 Spacer()
             }
             .ignoresSafeArea()
-             
+          
         } else {
-            ZStack {
                 ScrollView {
                     HStack(spacing: 14) {
                         Button {
@@ -121,31 +120,33 @@ struct ControlsView: View {
                     .padding(.bottom, 24)
                     Spacer()
                     LazyVGrid(columns: Array(repeating: GridItem(), count: 2), spacing: 33) {
-                        workoutDataItem(value: "\((workoutManager.distance/1000).formatted(.number.precision(.fractionLength(1))))", label: "킬로미터")
-                        workoutDataItem(value: "\(workoutManager.pace.formatted(.number.precision(.fractionLength(0))))", label: "평균 페이스")
-                        workoutDataItem(value: "\(workoutManager.activeEnergy.formatted(.number.precision(.fractionLength(0))))", label: "칼로리")
+                        workoutDataItem(value: "\((workoutManager.distance/1000).formatted(.number.precision(.fractionLength(2))))", label: "킬로미터")
+                        workoutDataItem(value: workoutManager.pace > 0
+                                        ? String(format: "%02d’%02d’’", Int(workoutManager.pace) / 60, Int(workoutManager.pace) % 60)
+                                        : "-’--’’",
+                                    label: "평균 페이스")
+                        workoutDataItem(value: "\(workoutManager.calorie.formatted(.number.precision(.fractionLength(0))))", label: "칼로리")
                         workoutDataItem(value: "\(workoutManager.averageHeartRate.formatted(.number.precision(.fractionLength(0))))", label: "BPM")
                     }
                     .padding(.bottom, 20)
                     Spacer()
                 }
-                VStack {
+                .navigationBarBackButtonHidden()
+                .overlay(alignment: .topLeading) {
                     Text("일시정지됨")
                         .bold()
                         .foregroundStyle(.green)
                         .padding()
                         .padding(.top, 5)
                         .padding(.leading, 20)
-                        .frame(maxWidth: .infinity, alignment: .topLeading)
+                        .frame(maxWidth: .infinity, alignment: .leading)
                         .background {
                             Rectangle()
                                 .foregroundStyle(.thinMaterial)
                         }
-                    Spacer()
+                        .ignoresSafeArea()
                 }
-                .ignoresSafeArea()
-            }
-           
+            
         }
     }
 }
@@ -155,13 +156,10 @@ extension ControlsView {
         VStack(spacing: 0) {
             Text(value)
                 .foregroundColor(.white)
-                .font(
-                    Font.custom("SF Pro Display", size: 21)
-                        .weight(.medium)
-                    )
+                .font(.system(size: 24).weight(.semibold))
                 .foregroundColor(.white)
             Text(label)
-                .font(Font.custom("SF Pro Display", size: 11))
+                .font(.system(size: 11))
                 .foregroundColor(Color.gray500)
         }
     }
